clip/pyproject.toml
[project]
name = "clip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.6",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "vecs>=0.4.5",
]
